Add timestamp_mask and hca_core_clock to ibv_query_device_ex
The fields timestamp_mask and hca_core_clock were added
to the extended version of ibv_query_device verb.
timestamp_mask represents the allowed mask of the timestamp.
Users could infer the accuracy of the reported possible
timestamp.
hca_core_clock represents the frequency of the HCA (in HZ).
Since timestamp and reading the HCA's core clock is given
in hardware cycles, knowing the frequency is mandatory in order
to convert this number into seconds.

issue: 418573
Change-Id: Ia79c0441da3be4df3d664e88e0883c4b729c2561
Signed-off-by: Matan Barak <matanb@mellanox.com>